Use built in ttl support for redis to make keys expire.
We were incorrectly setting the options for each cache region, now we
apply the defaults for each region or leave the override if one is
present for that region.

Also, previously we were passing all options when instantiating a store but now
we correctly get 'store_opts' and pass them in.