bump requests to the newest version
This is required because the older requests version we used
have a buggy urllib3 vendored in that breaks the py3
integration (live tests were failing).

More info can be found here:
https://github.com/stormpath/stormpath-sdk-python/issues/69

It's worth noting that we could not use older versions than 2.3.0 because of
another bug that was fixed in that release.

https://github.com/kennethreitz/requests/issues/1887

This was only when using pyopenssl, which is optional, the bug
was not triggered if request falls back to the stdlibs ssl
implementation. Still, it would have failed loudly is users
were using pyopenssl for any reason.

I've bumped to the latest version cause I see no benefit in using the
older versions seeing as 2.3.0 (which is the lowest version we can use)
already introduces breaking API changes and the subsequent releases have
a lot of important bugfixes (and no API changes, accoring to the
changelog). Still this change may break users code if they interact with
requests directly.