Add optional account store for password reset
When creating a password reset token the user can now specify an account
store that hold the desired account. If no account store is specified
the Backend uses the default account store to look for the account.

This is useful if one has multiple account store for an application
that hold the same email address.