Add exponential backoff strategy
If a request fails with status code 429 (backoff) or any 5xx error
we will backoff for a few ms and retry the request after that.

The user can supply a default 'get_delay' function to the http executor
which will be used to calculate the number of ms to wait before
retrying. This is done through the Client like so:

Client(backoff_strategy=lambda x: 0)

The function must take one parameter which is the number of
retries already done and return some number of milliseconds.

The maximum number of retries is 4 and is specified in the
DEFAULT_MAX_RETRIES constant on the HttpExecutor class.

The backoff is done using the sleep function from the time module.
This *should* work with gevent nicely because gevent pathces that
module. If this would actually logically work with the flow of the code
given that gevent uses implicit yielding and can suspend your code at
any time is an open question.

Even more so, the actual HTTP calls that
the requests package makes will still block even with gevent.
There are separate packages like grequests that should work if swaped
with the actual requests package. But then the actual library needs to
be retrofitted to work with this change. Another question is what if
a user is using another async lib? Like twisted or asyncio?