Changed assertion because of PyPy's differences.
In PyPy env, this:
    hrefs = [r.href for r in rl]
will call __len__ method, and then __iter__.
In Python envs, it will only call __iter__.
Because _ensure_data() is called in both __iter__ and __len__, for PyPy env it will call get_resource with params=None, and then with params={'limit': 2, 'offset': 2}.
The latter call doesn't happen when using Python.

So I decided to use assert_any_call() instead of assert_called_with(), because it passes if the mock has ever been called, unlike assert_called_with() that only pass if the call is the most recent one. This way it will work for both Python and Pypy.