Moved authentication details from data to params.
For some reason, live tests didn't caught this bug. But when tested in console, it would fail with "Authentication with a valid API Key is required." (as described in https://github.com/stormpath/stormpath-sdk-python/issues/202). I fixed this and added mock tests to make sure that params are passed correctly.