Error message as value for developerMessage key only when needed.
If error message is already a dict with developerMessage in it, there is no need to set it as value for another developerMessage key.
This caused weird behaviour described in https://github.com/stormpath/stormpath-sdk-python/issues/236 .
Old code used to pass dict with this format:
{
    'developerMessage': {
        'status': STATUS,
        'developerMessage': DEV_MSG,
        'message': USR_MSG,
        'code': CODE,
        'moreInfo': MORE_INFO
    }
}
in case proper dict was returned from Stormpath. This was introduced in https://github.com/stormpath/stormpath-sdk-python/commit/d741d2e815518bd99ba20885517025bea12f9366 while trying to fix the case when Error constructor doesn't get a dict.